fix(trial_balance): Calculate proper initial balance for last period on fiscal year when using in co
account_fiscal_year_closing module requieres two extra special periods at the end of the year for
hold the account moves that creates for close fiscal year. That extra special periods is causing
troubles getting the probler initial balances for last period in fiscal year. This is a temporal fix
for the problem while we figure what is the best option to work with: 1) Adapt modules to
fiscal_year_clossing or 2) Stop using special periods for hold account moves on fiscal year closing